+++ projects/kiwano/kiwano.vcxproj.filters
@@ -285,12 +285,6 @@
     <ClInclude Include="..\..\src\kiwano\core\Any.h">
       <Filter>core</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\kiwano\core\Json.h">
-      <Filter>core</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\src\kiwano\core\Xml.h">
-      <Filter>core</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\src\kiwano\core\IntrusiveList.h">
       <Filter>core</Filter>
     </ClInclude>
@@ -321,20 +315,26 @@
     <ClInclude Include="..\..\src\kiwano\render\FrameSequence.h">
       <Filter>render</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\kiwano\core\Task.h">
-      <Filter>core</Filter>
+    <ClInclude Include="..\..\src\kiwano\utils\EventTicker.h">
+      <Filter>utils</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\kiwano\core\TaskManager.h">
-      <Filter>core</Filter>
+    <ClInclude Include="..\..\src\kiwano\utils\Json.h">
+      <Filter>utils</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\kiwano\core\Timer.h">
-      <Filter>core</Filter>
+    <ClInclude Include="..\..\src\kiwano\utils\Task.h">
+      <Filter>utils</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\kiwano\core\Ticker.h">
-      <Filter>core</Filter>
+    <ClInclude Include="..\..\src\kiwano\utils\TaskManager.h">
+      <Filter>utils</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\kiwano\core\EventTicker.h">
-      <Filter>core</Filter>
+    <ClInclude Include="..\..\src\kiwano\utils\Ticker.h">
+      <Filter>utils</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\kiwano\utils\Timer.h">
+      <Filter>utils</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\kiwano\utils\Xml.h">
+      <Filter>utils</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
@@ -542,20 +542,20 @@
     <ClCompile Include="..\..\src\kiwano\render\FrameSequence.cpp">
       <Filter>render</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\kiwano\core\TaskManager.cpp">
-      <Filter>core</Filter>
+    <ClCompile Include="..\..\src\kiwano\utils\EventTicker.cpp">
+      <Filter>utils</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\kiwano\core\Task.cpp">
-      <Filter>core</Filter>
+    <ClCompile Include="..\..\src\kiwano\utils\Task.cpp">
+      <Filter>utils</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\kiwano\core\Timer.cpp">
-      <Filter>core</Filter>
+    <ClCompile Include="..\..\src\kiwano\utils\TaskManager.cpp">
+      <Filter>utils</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\kiwano\core\Ticker.cpp">
-      <Filter>core</Filter>
+    <ClCompile Include="..\..\src\kiwano\utils\Ticker.cpp">
+      <Filter>utils</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\kiwano\core\EventTicker.cpp">
-      <Filter>core</Filter>
+    <ClCompile Include="..\..\src\kiwano\utils\Timer.cpp">
+      <Filter>utils</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
