+++ src/kiwano-network/HttpRequest.h
@@ -21,7 +21,7 @@
 #pragma once
 #include <kiwano/core/Common.h>
 #include <kiwano/core/ObjectBase.h>
-#include <kiwano/core/Json.h>
+#include <kiwano/utils/Json.h>
 
 namespace kiwano
 {
