+++ src/kiwano/kiwano.h
@@ -52,11 +52,6 @@
 #include <kiwano/core/Resource.h>
 #include <kiwano/core/SmartPtr.hpp>
 #include <kiwano/core/Time.h>
-#include <kiwano/core/Timer.h>
-#include <kiwano/core/Ticker.h>
-#include <kiwano/core/EventTicker.h>
-#include <kiwano/core/Task.h>
-#include <kiwano/core/TaskManager.h>
 #include <kiwano/core/event/Event.h>
 #include <kiwano/core/event/KeyEvent.h>
 #include <kiwano/core/event/MouseEvent.h>
@@ -120,3 +115,8 @@
 #include <kiwano/utils/LocalStorage.h>
 #include <kiwano/utils/ResourceCache.h>
 #include <kiwano/utils/UserData.h>
+#include <kiwano/utils/Timer.h>
+#include <kiwano/utils/Ticker.h>
+#include <kiwano/utils/EventTicker.h>
+#include <kiwano/utils/Task.h>
+#include <kiwano/utils/TaskManager.h>
