+++ src/kiwano/utils/ResourceCache.h
@@ -19,11 +19,11 @@
 // THE SOFTWARE.
 
 #pragma once
+#include <kiwano/core/Resource.h>
+#include <kiwano/utils/Json.h>
+#include <kiwano/utils/Xml.h>
 #include <kiwano/render/Frame.h>
 #include <kiwano/render/FrameSequence.h>
-#include <kiwano/core/Resource.h>
-#include <kiwano/core/Json.h>
-#include <kiwano/core/Xml.h>
 #include <kiwano/render/Font.h>
 #include <kiwano/render/GifImage.h>
