+++ src/kiwano/utils/Task.cpp
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#include <kiwano/core/Task.h>
+#include <kiwano/utils/Task.h>
 
 namespace kiwano
 {
