+++ src/kiwano/utils/Ticker.cpp
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#include <kiwano/core/Ticker.h>
+#include <kiwano/utils/Ticker.h>
 
 namespace kiwano
 {
